client/ai_ops/requirements.txt
#requirements
asyncio
langchain-mcp-adapters
langchain-openai
langgraph
markitdown>=0.0.1a3
oracledb>=2.0.0
python-dotenv>=1.0.0
appdirs
oci
oracle_ads
langchain-core
langchain-oci==0.1.6
mcp>=1.11,<1.12 
fastmcp>=2.5,<2.6
uvicorn>=0.30.0 
anyio>=4.4,<5 
httpx>=0.27
plotly
matplotlib
geopandas
streamlit
streamlit-autorefresh
graphviz